+++ ddprof-lib/src/main/cpp/perfEvents_linux.cpp
@@ -147,7 +147,7 @@ static bool setPmuConfig(const char *device, const char *param, __u64 *config,
   ssize_t r = read(fd, buf, sizeof(buf));
   close(fd);
 
-  if (r > 0 && r < sizeof(buf)) {
+  if (r > 0 && r < (int)sizeof(buf)) {
     if (strncmp(buf, "config:", 7) == 0) {
       config[0] |= val << atoi(buf + 7);
       return true;
